refactor: extract shared views infrastructure into views.go
Move package-level vars, runOnMainThread, ToastAdder interface,
UserHome struct, New constructor, and helper functions (updateBadgeCount,
GetPage, createPage) out of userhome.go into views.go. This prepares
for splitting per-page code into separate files.
